+++ pip-services4-aws-dart/lib/src/container/CommandableLambdaFunction.dart
@@ -75,7 +75,9 @@ abstract class CommandableLambdaFunction extends LambdaFunction {
         dynamic result;
         try {
           result = await command.execute(context, args);
+          timing.endTiming();
         } catch (ex) {
+          timing.endTiming();
           rethrow;
         } finally {
           timing.endTiming();