+++ pip-services4-azure-dart/lib/pip_services4_azure.dart
@@ -1,0 +1,6 @@
+library pip_services4_azure;
+
+export './src/clients/clients.dart';
+export './src/container/container.dart';
+export './src/connect/connect.dart';
+export './src/controllers/controllers.dart';