+++ pip-services4-azure-dart/lib/src/clients/clients.dart
@@ -1,0 +1,2 @@
+export 'AzureFunctionClient.dart';
+export 'CommandableAzureFunctionClient.dart';