+++ pip-services4-azure-dart/lib/src/connect/connect.dart
@@ -1,0 +1,2 @@
+export 'AzureFunctionConnectionParams.dart';
+export 'AzureFunctionConnectionResolver.dart';