+++ pip-services4-azure-dart/test/controllers/DummyAzureFunction.dart
@@ -1,0 +1,13 @@
+import 'package:pip_services4_azure/pip_services4_azure.dart';
+import '../DummyFactory.dart';
+
+class DummyAzureFunction extends AzureFunction {
+  DummyAzureFunction() : super('dummy', 'Dummy Azure function') {
+    factories.add(DummyFactory());
+  }
+
+  @override
+  void register() {
+    // TODO: implement register
+  }
+}